Support for dbt-seed, dbt-snapshot and dbt-build
