+++ dbt_checkpoint/dbt_seed.py
@@ -4,7 +4,7 @@ from dbt_checkpoint.dbt_models_command import parse_and_run
 
 
 def main(argv: Optional[Sequence[str]] = None) -> int:
-    return parse_and_run("run", argv)
+    return parse_and_run("seed", argv)
 
 
 if __name__ == "__main__":
