+++ setup.cfg
@@ -60,11 +60,14 @@ console_scripts =
     check-source-has-tests-by-type = dbt_checkpoint.check_source_has_tests_by_type:main
     check-source-has-tests = dbt_checkpoint.check_source_has_tests:main
     check-source-tags = dbt_checkpoint.check_source_tags:main
+    dbt-build = dbt_checkpoint.dbt_build:main
     dbt-clean = dbt_checkpoint.dbt_clean:main
     dbt-compile = dbt_checkpoint.dbt_compile:main
     dbt-deps = dbt_checkpoint.dbt_deps:main
     dbt-docs-generate = dbt_checkpoint.dbt_docs_generate:main
     dbt-run = dbt_checkpoint.dbt_run:main
+    dbt-seed = dbt_checkpoint.dbt_seed:main
+    dbt-snapshot = dbt_checkpoint.dbt_snapshot:main
     dbt-test = dbt_checkpoint.dbt_test:main
     generate-missing-sources = dbt_checkpoint.generate_missing_sources:main
     generate-model-properties-file = dbt_checkpoint.generate_model_properties_file:main
